Gestational DNAm age estimation using different DNA methylation clocks.

Gestational DNAm age estimation using different DNA methylation clocks.

Usage

DNAmGA(
  x,
  toBetas = FALSE,
  fastImp = FALSE,
  normalize = FALSE,
  age,
  cell.count = TRUE,
  cell.count.reference = "andrews and bakulski cord blood",
  min.perc = 0.8,
  ...
)

Arguments

x

data.frame (Individual in columns, CpGs in rows, CpG names in first colum - i.e. Horvath's format), matrix (individuals in columns and Cpgs in rows having CpG names in the rownames), ExpressionSet or GenomicRatioSet.

toBetas

Should data be transformed to beta values? Default is FALSE. If TRUE, it implies data are M values.

fastImp

Is fast imputation performed if necessary? (see details). Default is FALSE

normalize

Is Horvath's normalization performed? By default is FALSE

age

individual's chronological age. Required to compute gestational age difference output

cell.count

Are cell counts estimated? Default is TRUE.

cell.count.reference

Used when 'cell.count' is TRUE. Default is "blood gse35069 complete". See 'meffil::meffil.list.cell.count.references()' for possible values.

min.perc

Indicates the minimum conicidence percentage required between CpGs in or dataframee x and CpGs in clock coefficients to perform the calculation. If min.prec is too low, the estimated gestational DNAm age can be poor

...

Other arguments to be passed through impute package

Value

the estimated gestational DNAm age

Details

Imputation is performed when having missing data. Fast imputation is performed by ... what about imputing only when CpGs for the clock are missing?

Examples

TestDataset <- get_TestDataset()
#> snapshotDate(): 2022-04-26
#> see ?methylclockData and browseVignettes('methylclockData') for documentation
#> loading from cache
TestDataset[1:5, ]
#>      CpGName    Sample1    Sample2    Sample3
#> 1 cg00000292 0.72546496 0.72350947 0.69023377
#> 2 cg00002426 0.85091763 0.80077888 0.80385777
#> 3 cg00003994 0.05125853 0.05943935 0.05559333
#> 4 cg00005847 0.08775420 0.11722333 0.10845113
#> 5 cg00006414 0.03982478 0.06146891 0.03491992
ga.test <- DNAmGA(TestDataset)
#> Warning: CpGs in all Gestational Age clocks are not present in your
#>         data. Try 'checkClocksGA' function to find the missing CpGs of
#>                 each method.
#> Warning: The number of missing CpGs forBohlinclock exceeds 80%.
#>   ---> This DNAm clock will be NA.
#> Warning: The number of missing CpGs forEPICclock exceeds 80%.
#>   ---> This DNAm clock will be NA.
#> Warning: The number of missing CpGs for Lee clocks exceeds 80%.
#>                 
#>   ---> This DNAm clock will be NA.