Get common CpGs

Show the required CpGs contained on input data for the implemented clocks

Usage

commonClockCpgs(object, clock)

Arguments

object

resulting object from checkClocks functions

clock

string with the implemented clock, possible values are : "Knight", "Bohlin", "Mayne" and "Lee", "Horvath", "Hannum", "Levine", "skinHorvath", "PedBE", "Wu" and "TL"

Value

The common CpGs between input data and defined GA clock

Examples

TestDataset <- get_TestDataset()
#> snapshotDate(): 2022-04-26
#> see ?methylclockData and browseVignettes('methylclockData') for documentation
#> loading from cache
cpgs.missing.GA <- checkClocksGA(TestDataset)
#> There are some clocks that cannot be computed since
#>                 your data donot contain the required CpGs.These are the total
#>                 number of missing CpGs for each clock : 
#>  
#>    clock Cpgs_in_clock missing_CpGs percentage
#> 1 Knight           148            0        0.0
#> 2 Bohlin            87           87      100.0
#> 3  Mayne            62            0        0.0
#> 4    Lee          1125         1072       95.3
#> 5   EPIC           176          170       96.6
cpgs.missing <- checkClocks(TestDataset)
#> There are some clocks that cannot be computed since your 
#>         data do not contain the required CpGs.These are the total number 
#>         of missing CpGs for each clock : 
#>  
#>         clock Cpgs_in_clock missing_CpGs percentage
#> 1     Horvath           353            2        0.6
#> 2      Hannum            71           64       90.1
#> 3      Levine           513            3        0.6
#> 4 SkinHorvath           391          283       72.4
#> 5       PedBE            94           91       96.8
#> 6          Wu           111            2        1.8
#> 7          TL           140          137       97.9
#> 8        BLUP        319607       300288       94.0
#> 9          EN           514          476       92.6
commonClockCpgs(cpgs.missing.GA, "Bohlin")
#>  [1] "cg00153101" "cg00602416" "cg00711496" "cg01190109" "cg01635555"
#>  [6] "cg01833485" "cg02324006" "cg02405476" "cg02567958" "cg02642822"
#> [11] "cg03108070" "cg03281561" "cg03337084" "cg03507326" "cg03710860"
#> [16] "cg03729251" "cg03773820" "cg03963689" "cg04347477" "cg04685228"
#> [21] "cg05053327" "cg05544807" "cg05877497" "cg06753281" "cg06897661"
#> [26] "cg07106169" "cg07676709" "cg07738730" "cg07749613" "cg07788865"
#> [31] "cg07835443" "cg08326019" "cg08620426" "cg08943494" "cg09447786"
#> [36] "cg10308785" "cg11124260" "cg11294761" "cg11864574" "cg12880227"
#> [41] "cg12999267" "cg13036381" "cg13066703" "cg13433246" "cg13641317"
#> [46] "cg13733403" "cg13959344" "cg13982823" "cg14276580" "cg14427590"
#> [51] "cg15035133" "cg15131146" "cg15165154" "cg15908709" "cg16187883"
#> [56] "cg16348385" "cg17022232" "cg18183624" "cg18217136" "cg18954401"
#> [61] "cg19057830" "cg19439123" "cg19875532" "cg20301308" "cg20303561"
#> [66] "cg20816447" "cg21081878" "cg21143441" "cg21155834" "cg21221899"
#> [71] "cg21707172" "cg21878650" "cg22761205" "cg22796593" "cg22797644"
#> [76] "cg23051248" "cg23346945" "cg23403099" "cg23457357" "cg24041556"
#> [81] "cg24087613" "cg24366564" "cg25150953" "cg25531857" "cg25639749"
#> [86] "cg26077811" "cg26092675"
commonClockCpgs(cpgs.missing, "Hannum")
#>  [1] "cg20822990"      "cg22512670"      "cg25410668"      "cg04400972"     
#>  [5] "cg16054275"      "cg10501210"      "ch.2.30415474F"  "cg22158769"     
#>  [9] "cg02085953"      "cg06639320"      "cg22454769"      "cg24079702"     
#> [13] "cg23606718"      "cg22016779"      "cg03607117"      "cg07553761"     
#> [17] "cg00481951"      "cg25478614"      "cg25428494"      "cg02650266"     
#> [21] "cg08234504"      "cg23500537"      "cg20052760"      "cg16867657"     
#> [25] "cg06685111"      "cg00486113"      "cg13001142"      "cg20426994"     
#> [29] "cg14361627"      "cg08097417"      "cg07955995"      "cg22285878"     
#> [33] "cg03473532"      "cg08540945"      "cg07927379"      "cg16419235"     
#> [37] "cg07583137"      "cg22796704"      "cg19935065"      "cg23091758"     
#> [41] "cg23744638"      "cg04940570"      "cg11067179"      "cg22213242"     
#> [45] "cg06419846"      "cg02046143"      "cg00748589"      "cg18473521"     
#> [49] "cg01528542"      "ch.13.39564907R" "cg03032497"      "cg04875128"     
#> [53] "cg09651136"      "cg03399905"      "cg04416734"      "cg07082267"     
#> [57] "cg14692377"      "cg06874016"      "cg21139312"      "cg02867102"     
#> [61] "cg19283806"      "cg14556683"      "cg07547549"      "cg08415592"