Estimate cell counts for a beta matrix from a reference

Estimate cell type ratios from methylation profiles of purified cell populations (Infinium HumanMethylation450 BeadChip).

Usage

meffilEstimateCellCountsFromBetas(beta, cellTypeReference, verbose = FALSE)

Arguments

beta

Matrix of Illumina 450K methylation levels (rows = CpG sites, columns = subjects).

cellTypeReference

Character string name of the cell type reference to use for estimating cell counts. See meffilListCellTypeReferences() for a list of available references. New references can be created using

verbose

If TRUE, then status messages are printed during execution (Default: FALSE).

Value

A matrix of cell count estimates.

Results should be nearly identical to minfi::estimateCellCounts()

betas

Details

ORIGINAL AUTHOR: Matthew Suderman The original meffil.list.cellTypeReferences and get.cellTypeReference function from meffil v1.0.0 downloaded from githug : https://github.com/perishky/meffil

Examples

cell.count.reference <- "andrews and bakulski cord blood"
TestDataset <- get_TestDataset()
#> snapshotDate(): 2022-04-26
#> see ?methylclockData and browseVignettes('methylclockData') for documentation
#> loading from cache
cpgs <- t(as.matrix(TestDataset[, -1]))
colnames(cpgs) <- TestDataset$CpGName
meffilEstimateCellCountsFromBetas(t(cpgs), cell.count.reference)
#>             Bcell      CD4T         CD8T       Gran         Mono        NK
#> Sample1 0.4572106 0.3654086 6.035999e-18 0.00000000 1.298103e-18 0.4352757
#> Sample2 0.5125346 0.4289509 1.387803e-17 0.04758639 2.561457e-02 0.3289237
#> Sample3 0.4471258 0.2667828 0.000000e+00 0.04658645 4.674347e-02 0.3730971
#>              nRBC
#> Sample1 0.1006619
#> Sample2 0.0000000
#> Sample3 0.1466445